Use munit in MinSuite
